# api.yaml
openapi: 3.0.3
info:
  title: Topic and Student Management API
  version: 1.0.0
  description: API documentation for managing educational topics, subtopics, students, and system health.
paths:
  /api/topics:
    get:
      summary: Get all topics
      responses:
        '200':
          description: A list of topics retrieved successfully.
    post:
      summary: Create a new topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicInput'
      responses:
        '201':
          description: Topic created successfully.

  /api/topics/{topicId}:
    parameters:
      - name: topicId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a topic by ID
      responses:
        '200':
          description: Topic details retrieved successfully.
        '404':
          description: Topic not found.
    put:
      summary: Update an entire topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicInput'
      responses:
        '200':
          description: Topic updated successfully.
        '404':
          description: Topic not found.
    delete:
      summary: Delete a topic
      responses:
        '204':
          description: Topic deleted successfully.
        '404':
          description: Topic not found.

  /api/topics/{topicId}/position:
    parameters:
      - name: topicId
        in: path
        required: true
        schema:
          type: string
    patch:
      summary: Partially update topic position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: integer
              required:
                - position
      responses:
        '200':
          description: Topic position updated successfully.
        '404':
          description: Topic not found.

  /api/topics/{topicId}/subtopics:
    parameters:
      - name: topicId
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Create a subtopic under a specific topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubtopicInput'
      responses:
        '201':
          description: Subtopic created successfully.
        '404':
          description: Parent topic not found.

  /api/topics/{topicId}/subtopics/{subtopicId}:
    parameters:
      - name: topicId
        in: path
        required: true
        schema:
          type: string
      - name: subtopicId
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Update a specific subtopic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubtopicInput'
      responses:
        '200':
          description: Subtopic updated successfully.
        '404':
          description: Topic or subtopic not found.
    delete:
      summary: Delete a specific subtopic
      responses:
        '204':
          description: Subtopic deleted successfully.
        '404':
          description: Topic or subtopic not found.

  /api/students:
    get:
      summary: Get students with optional search filtering
      parameters:
        - name: query
          in: query
          required: false
          description: Search term to filter students by name, email, or other properties.
          schema:
            type: string
      responses:
        '200':
          description: A list of students retrieved successfully.
    post:
      summary: Create a new student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student profile created successfully.

  /api/students/{studentId}:
    parameters:
      - name: studentId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get a student profile by ID
      responses:
        '200':
          description: Student profile details retrieved successfully.
        '404':
          description: Student not found.
    put:
      summary: Update an entire student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student profile updated successfully.
        '404':
          description: Student not found.
    delete:
      summary: Delete a student record
      responses:
        '204':
          description: Student profile deleted successfully.
        '404':
          description: Student not found.

  /api/health:
    get:
      summary: Check system health status
      responses:
        '200':
          description: System is up and healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: UP

components:
  schemas:
    TopicInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
    SubtopicInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
      required:
        - title
    StudentInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
      required:
        - firstName
        - lastName
        - email